requirements.txt
fastapi
uvicorn
diffusers
transformers
torch
pillow
streamlit
requests
pillow
